reimplement derived token addresses based on go library

+++ solana-kotlin/src/commonMain/kotlin/net/avianlabs/solana/domain/program/ProgramDerivedAddress.kt
@@ -4,4 +4,4 @@ import kotlinx.serialization.Serializable
 import net.avianlabs.solana.domain.core.PublicKey
 
 @Serializable
-public data class ProgramDerivedAddress(val address: PublicKey, val nonce: Int)
+public data class ProgramDerivedAddress(val address: PublicKey, val nonce: UByte)
